incident_report/requirements.txt
Django==3.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
numpy==1.21.0

